# perf: parse CbTx only once perf block